Renamed junit test according to Java conventions
